borked radix sort

+++ src/BarPlugin.rs
@@ -4,14 +4,14 @@ use macroquad::rand;
 
 #[derive(Debug, Clone, Copy)]
 pub struct Bar {
-    pub position:i32,
+    pub position:usize,
     pub color:color::Color
 }
 
 
 
 impl Bar{
-    pub fn new(position:i32, hsl_color:f32) -> Self{
+    pub fn new(position:usize, hsl_color:f32) -> Self{
         Bar{
             position,
             color: color::hsl_to_rgb((hsl_color as f32) , 1.0, 0.5),
